+++ pyuic.json
@@ -21,6 +21,10 @@
             "randovania/games/dread/gui/generated"
         ],
         [
+            "randovania/games/factorio/gui/ui_files/*.ui",
+            "randovania/games/factorio/gui/generated"
+        ],
+        [
             "randovania/games/fusion/gui/ui_files/*.ui",
             "randovania/games/fusion/gui/generated"
         ],
