+++ randovania/games/factorio/gui/ui_files/.gitignore
@@ -1,0 +1,1 @@
+*.ui.autosave
